# Avoid of alloc for func args copy