--- a/expr/functions/averageOutsidePercentile/function.go
+++ b/expr/functions/averageOutsidePercentile/function.go
@@ -33,7 +33,7 @@ func (f *averageOutsidePercentile) Do(ctx context.Context, e parser.Expr, from, 
 		return nil, parser.ErrMissingArgument
 	}
 
-	args, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	args, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}