--- a/expr/functions/exponentialMovingAverage/function_test.go
+++ b/expr/functions/exponentialMovingAverage/function_test.go
@@ -1,6 +1,7 @@
 package exponentialMovingAverage
 
 import (
+	"context"
 	"testing"
 
 	"github.com/go-graphite/carbonapi/expr/helper"
@@ -34,6 +35,15 @@ func TestExponentialMovingAverage(t *testing.T) {
 			},
 		},
 		{
+			"exponentialMovingAverage(metric1,'3s')",
+			map[parser.MetricRequest][]*types.MetricData{
+				{"metric1", 0, 1}: {types.MakeMetricData("metric1", []float64{2, 4, 6, 8, 12, 14, 16, 18, 20}, 1, startTime)},
+			},
+			[]*types.MetricData{
+				types.MakeMetricData("exponentialMovingAverage(metric1,\"3s\")", []float64{4, 6, 9, 11.5, 13.75, 15.875, 17.9375}, 1, 0).SetTag("exponentialMovingAverage", "3"),
+			},
+		},
+		{
 			// if the window is larger than the length of the values, the result should just be the average.
 			// this matches graphiteweb's behavior
 			"exponentialMovingAverage(metric1,100)",
@@ -52,5 +62,49 @@ func TestExponentialMovingAverage(t *testing.T) {
 		t.Run(testName, func(t *testing.T) {
 			th.TestEvalExpr(t, &tt)
 		})
+	}
+}
+
+func BenchmarkExponentialMovingAverage(b *testing.B) {
+	target := "exponentialMovingAverage(metric1,3)"
+	metrics := map[parser.MetricRequest][]*types.MetricData{
+		{"metric[1234]", 0, 1}: {types.MakeMetricData("metric1", []float64{2, 4, 6, 8, 12, 14, 16, 18, 20}, 1, 0)},
+	}
+
+	evaluator := metadata.GetEvaluator()
+	exp, _, err := parser.ParseExpr(target)
+	if err != nil {
+		b.Fatalf("failed to parse %s: %+v", target, err)
+	}
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		g, err := evaluator.Eval(context.Background(), exp, 0, 1, metrics)
+		if err != nil {
+			b.Fatalf("failed to eval %s: %+v", target, err)
+		}
+		_ = g
+	}
+}
+
+func BenchmarkExponentialMovingAverageStr(b *testing.B) {
+	target := "exponentialMovingAverage(metric1,'3s')"
+	metrics := map[parser.MetricRequest][]*types.MetricData{
+		{"metric[1234]", 0, 1}: {types.MakeMetricData("metric1", []float64{2, 4, 6, 8, 12, 14, 16, 18, 20}, 1, 0)},
+	}
+
+	evaluator := metadata.GetEvaluator()
+	exp, _, err := parser.ParseExpr(target)
+	if err != nil {
+		b.Fatalf("failed to parse %s: %+v", target, err)
+	}
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		g, err := evaluator.Eval(context.Background(), exp, 0, 1, metrics)
+		if err != nil {
+			b.Fatalf("failed to eval %s: %+v", target, err)
+		}
+		_ = g
 	}
 }