--- a/expr/functions/holtWintersConfidenceArea/function_cairo.go
+++ b/expr/functions/holtWintersConfidenceArea/function_cairo.go
@@ -39,7 +39,7 @@ func (f *holtWintersConfidenceArea) Do(ctx context.Context, e parser.Expr, from,
 		return nil, err
 	}
 
-	args, err := helper.GetSeriesArg(ctx, e.Args()[0], from-bootstrapInterval, until, values)
+	args, err := helper.GetSeriesArg(ctx, e.Arg(0), from-bootstrapInterval, until, values)
 	if err != nil {
 		return nil, err
 	}