--- a/expr/functions/logit/function.go
+++ b/expr/functions/logit/function.go
@@ -32,7 +32,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 // logarithm(seriesList, base=10)
 // Alias: log
 func (f *logit) Do(ctx context.Context, e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}