--- a/expr/functions/removeBetweenPercentile/function.go
+++ b/expr/functions/removeBetweenPercentile/function.go
@@ -36,7 +36,7 @@ func (f *removeBetweenPercentile) Do(ctx context.Context, e parser.Expr, from, u
 		return nil, parser.ErrMissingArgument
 	}
 
-	args, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	args, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}