--- a/expr/functions/removeEmptySeries/function.go
+++ b/expr/functions/removeEmptySeries/function.go
@@ -41,7 +41,7 @@ func (f *removeEmptySeries) Do(ctx context.Context, e parser.Expr, from, until i
 		return []*types.MetricData{}, nil
 	}
 
-	if len(e.Args()) == 2 {
+	if e.ArgsLen() == 2 {
 		xFilesFactor, err = e.GetFloatArgDefault(1, float64(args[0].XFilesFactor)) // If set by setXFilesFactor, all series in a list will have the same value
 		if err != nil {
 			return nil, err