--- a/expr/functions/round/function.go
+++ b/expr/functions/round/function.go
@@ -30,7 +30,7 @@ func New(_ string) []interfaces.FunctionMetadata {
 
 // round(seriesList,precision)
 func (f *round) Do(ctx context.Context, e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}