--- a/expr/functions/scaleToSeconds/function.go
+++ b/expr/functions/scaleToSeconds/function.go
@@ -33,8 +33,8 @@ func (f *scaleToSeconds) Do(ctx context.Context, e parser.Expr, from, until int6
 	if e.ArgsLen() < 2 {
 		return nil, parser.ErrMissingArgument
 	}
-	
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}