--- a/expr/functions/sigmoid/function.go
+++ b/expr/functions/sigmoid/function.go
@@ -30,7 +30,7 @@ func New(_ string) []interfaces.FunctionMetadata {
 
 // sigmoid(seriesList)
 func (f *sigmoid) Do(ctx context.Context, e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}