--- a/expr/functions/sortBy/function.go
+++ b/expr/functions/sortBy/function.go
@@ -32,7 +32,7 @@ func New(configFile string) []interfaces.FunctionMetadata {
 
 // sortByMaxima(seriesList), sortByMinima(seriesList), sortByTotal(seriesList)
 func (f *sortBy) Do(ctx context.Context, e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
-	original, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	original, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}