--- a/expr/functions/stdev/function.go
+++ b/expr/functions/stdev/function.go
@@ -33,7 +33,7 @@ func (f *stdev) Do(ctx context.Context, e parser.Expr, from, until int64, values
 		return nil, parser.ErrMissingArgument
 	}
 
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}