--- a/expr/functions/timeSlice/function.go
+++ b/expr/functions/timeSlice/function.go
@@ -2,12 +2,13 @@ package timeSlice
 
 import (
 	"context"
+	"math"
+	"strconv"
+
 	"github.com/go-graphite/carbonapi/expr/helper"
 	"github.com/go-graphite/carbonapi/expr/interfaces"
 	"github.com/go-graphite/carbonapi/expr/types"
 	"github.com/go-graphite/carbonapi/pkg/parser"
-	"math"
-	"strconv"
 )
 
 type timeSlice struct {
@@ -47,7 +48,7 @@ func (f *timeSlice) Do(ctx context.Context, e parser.Expr, from, until int64, va
 	startStr := strconv.FormatInt(start, 10)
 	endStr := strconv.FormatInt(end, 10)
 
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}