--- a/expr/functions/unique/function.go
+++ b/expr/functions/unique/function.go
@@ -29,7 +29,7 @@ func New(_ string) []interfaces.FunctionMetadata {
 
 // unique(seriesList)
 func (f *unique) Do(ctx context.Context, e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) ([]*types.MetricData, error) {
-	arg, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	arg, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return nil, err
 	}