--- a/expr/rewrite/applyByNode/function.go
+++ b/expr/rewrite/applyByNode/function.go
@@ -29,7 +29,7 @@ func New(configFile string) []interfaces.RewriteFunctionMetadata {
 }
 
 func (f *applyByNode) Do(ctx context.Context, e parser.Expr, from, until int64, values map[parser.MetricRequest][]*types.MetricData) (bool, []string, error) {
-	args, err := helper.GetSeriesArg(ctx, e.Args()[0], from, until, values)
+	args, err := helper.GetSeriesArg(ctx, e.Arg(0), from, until, values)
 	if err != nil {
 		return false, nil, err
 	}