Cleanup of use e.Args()[N] and len(e.Args())

For example
```
$ go test -benchmem -run=^$ -bench ^BenchmarkExponentialMovingAverage github.com/go-graphite/carbonapi/expr/functions/exponentialMovingAverage

$ benchcmp exponentialMovingAverage1.txt exponentialMovingAverage2.txt 

benchmark                                  old ns/op     new ns/op     delta
BenchmarkExponentialMovingAverage-6        216           48.5          -77.58%
BenchmarkExponentialMovingAverageStr-6     436           134           -69.30%

benchmark                                  old allocs     new allocs     delta
BenchmarkExponentialMovingAverage-6        3              0              -100.00%
BenchmarkExponentialMovingAverageStr-6     6              2              -66.67%

benchmark                                  old bytes     new bytes     delta
BenchmarkExponentialMovingAverage-6        96            0             -100.00%
BenchmarkExponentialMovingAverageStr-6     148           16            -89.19%


```